+++ include/core/Scene.hpp
@@ -1,35 +1,40 @@
 #ifndef TNT_SCENE_HPP
 #define TNT_SCENE_HPP
 
-#include "Camera.hpp"
-#include "core/Space.hpp"
-#include "fileIO/AssetManager.hpp"
+#include <vector>
 
 // TODO: implement cutscenes (scenes where player input is ignored).
 // TODO: store a Quadtree on each Scene (or Space).
 // TODO: add more functionality.
+// TODO: move loader functions on class Game.
 
 // TODO(maybe):
-// this should be an interface.
+// make LoadAssets() a coroutine ??
 
 namespace tnt
 {
-    class Scene
-    {
-      public:
-        Scene();
-        ~Scene() noexcept;
+class Camera;
+class Window;
+class Space;
 
-        virtual void LoadAssets() = 0;
-        void Draw();
-        void Update();
+class Scene
+{
+public:
+    explicit Scene(std::shared_ptr<Window> window_);
+    ~Scene() noexcept;
 
-      private:
-        Camera *camera;
-        std::weak_ptr<Scene>
-            scenes; // TODO(maybe): THIS should be moved into game
-        AssetManager &assets;
-    };
+    // NOTE: fonts are loaded using LoadFont().
+    void LoadAssets(std::vector<std::string> const &assets);
+    void LoadFont(std::string_view name, int size);
+
+    void Draw();
+    void Update(long long time_);
+
+private:
+    Camera *camera;
+    std::weak_ptr<Window> window;
+    std::vector<Space *> spaces;
+};
 } // namespace tnt
 
 #endif //! TNT_SCENE_HPP