+++ src/core/Camera.cpp
@@ -1,22 +1,40 @@
-// This is an independent project of an individual developer. Dear PVS-Studio,
-// please check it. PVS-Studio Static Code Analyzer for C, C++, C#, and Java:
-// http://www.viva64.com
+// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
-#include "Camera.hpp"
+#include "core/Camera.hpp"
+#include "pcg/Random.hpp"
 
 ////////////
 // Camera //
 ////////////
 
 tnt::Camera::Camera(float x, float y, float w, float h)
-    : bounds{SDL_FRect{x, y, w, h}}
-{}
+    : bounds{x, y, w, h}, shaking{false}, tmpX{0.f}, tmpY{0.f} {}
 
-SDL_FRect tnt::Camera::FBounds() const noexcept { return SDL_FRect{bounds}; }
+tnt::Rectangle tnt::Camera::Bounds() const noexcept { return bounds; }
 
-SDL_Rect tnt::Camera::Bounds() const noexcept
+void tnt::Camera::Shake(long long time, float intensity) noexcept
 {
-    return SDL_Rect{(int)bounds.x, (int)bounds.y, (int)bounds.w, (int)bounds.h};
+    if (intensity > .1f)
+    {
+        if (!shaking)
+        {
+            tmpX = bounds.x;
+            tmpY = bounds.y;
+        }
+        else
+            shaking = true;
+        intensity = intensity * 0.9f;
+        float angle{randomFloat(0.f, 2 * PI)}; // random angle betwen [0, 360] degree
+        float xOffset{bounds.h * std::cosf(angle)};
+        float yOffset{bounds.w * std::sinf(angle)};
+        bounds.x = bounds.x + xOffset;
+        bounds.y = bounds.y + yOffset;
+        return;
+    }
+    shaking = false;
+    bounds.x = tmpX;
+    bounds.y = tmpY;
 }
 
 ///////////////////////
@@ -24,26 +42,25 @@ SDL_Rect tnt::Camera::Bounds() const noexcept
 ///////////////////////
 
 tnt::HorizontalCamera::HorizontalCamera(float x, float y, float w, float h)
-    : Camera{x, y, w, h}
-{}
+    : Camera{x, y, w, h} {}
 
 void tnt::HorizontalCamera::Move(float x, float) { bounds.x = bounds.x + x; }
 
-void tnt::HorizontalCamera::Move(tnt::Vector const& v)
+void tnt::HorizontalCamera::Move(tnt::Vector const &v)
 {
     bounds.x = bounds.x + v.x;
 }
 
 void tnt::HorizontalCamera::MoveTo(float x, float) { bounds.x = x; }
 
-void tnt::HorizontalCamera::MoveTo(tnt::Vector const& v) { bounds.x = v.x; }
+void tnt::HorizontalCamera::MoveTo(tnt::Vector const &v) { bounds.x = v.x; }
 
 void tnt::HorizontalCamera::CenterTo(float x, float)
 {
     bounds.x = (x - (bounds.w / 2));
 }
 
-void tnt::HorizontalCamera::CenterTo(tnt::Vector const& v)
+void tnt::HorizontalCamera::CenterTo(tnt::Vector const &v)
 {
     bounds.x = (v.x - (bounds.w / 2));
 }
@@ -53,8 +70,7 @@ void tnt::HorizontalCamera::CenterTo(tnt::Vector const& v)
 ///////////////////////////
 
 tnt::FullTrackingCamera::FullTrackingCamera(float x, float y, float w, float h)
-    : Camera{x, y, w, h}
-{}
+    : Camera{x, y, w, h} {}
 
 void tnt::FullTrackingCamera::Move(float x, float y)
 {
@@ -62,7 +78,7 @@ void tnt::FullTrackingCamera::Move(float x, float y)
     bounds.y = bounds.y + y;
 }
 
-void tnt::FullTrackingCamera::Move(Vector const& v)
+void tnt::FullTrackingCamera::Move(Vector const &v)
 {
     bounds.x = bounds.x + v.x;
     bounds.y = bounds.y + v.y;
@@ -78,7 +94,7 @@ void tnt::FullTrackingCamera::MoveTo(float x, float y)
     bounds.y = y;
 }
 
-void tnt::FullTrackingCamera::MoveTo(Vector const& v)
+void tnt::FullTrackingCamera::MoveTo(Vector const &v)
 {
     bounds.x = v.x;
     bounds.y = v.y;
@@ -90,7 +106,7 @@ void tnt::FullTrackingCamera::CenterTo(float x, float y)
     bounds.y = y - (bounds.h / 2);
 }
 
-void tnt::FullTrackingCamera::CenterTo(Vector const& v)
+void tnt::FullTrackingCamera::CenterTo(Vector const &v)
 {
     bounds.x = v.x - (bounds.w / 2);
     bounds.y = v.y - (bounds.h / 2);