+++ src/pcg/Random.hpp
@@ -8,6 +8,12 @@
 namespace tnt
 {
 
+float randomFloat(float min_, float max_) noexcept
+{
+    static std::default_random_engine eng;
+    static std::uniform_real_distribution<float> dist{min_, std::nextafter(max_, std::numeric_limits<float>::max())};
+    return dist(eng);
+}
 int randomValue(int min, int max) noexcept
 {
     std::random_device device;