+++ test/rpp/rpptest.cpp
@@ -1,56 +1,24 @@
 #include <iostream>
-
 #include <SDL2/SDL.h>
+#include "testlib.h"
 #include "exp/Runtimer.hpp"
 
-
-typedef int(*answer_t)();
-answer_t answer = NULL;
-
-int main([[maybe_unused]] int argc,[[maybe_unused]] char* argv[])
+int main([[maybe_unused]] int argc, [[maybe_unused]] char *argv[])
 {
-    using tnt::rpp::RuntimeManager;
+    using namespace tnt::rpp;
 
-	if(SDL_Init(SDL_INIT_EVERYTHING)<0)
-	{
-		std::cout << "couldn't init sdl2!! Error: " << SDL_GetError() << '\n';
-		return -1;
-	}
-	
-	RuntimeManager* rpp{new RuntimeManager{}};
+    RuntimeObject *object;
+    object->binPath = "test.dll";
+    object->procsToLoad.push_back("answer");
 
-    rpp->LoadObject("test", "testlib.c");
-	
-	std::cout << "test error " << SDL_GetError() << '\n';
-	
-	answer = *static_cast<answer_t*>(rpp->LoadFunction("test", "answer"));
-	
-	bool running{true};
-	
-	while(running)
-	{
-		rpp->Update();
-		
-		
-    if (answer != NULL)
-            std::cout << answer();
-        else
-		{
-			std::cout << "int answer() couldn't be loaded correctly! Error:" << SDL_GetError() 
-				<<".\n exiting...";
-			running = false;
-		}
-	}
-//        if (answer() == 43)
-//        {
-//            tnt::logger::info("Succesfully exiting...");
-//            running = false;
-//        }
-//        else
-//            continue;
-	
-	delete rpp;
-	SDL_Quit();
-	
+    while (true)
+    {
+        tnt::rpp::Update(object);
+        auto live_code{(decltype(answer) *)(object->processes["answer"])};
+        if (live_code)
+            live_code();
+        SDL_Delay(1000);
+    }
+
     return 0;
 }